+++ apps/app4/kernels/vector_scalar_mul.cc
@@ -15,7 +15,7 @@
 
 extern "C" {
 
-void vector_scalar_mul_aie_scalar(int32_t *a, int32_t *c, int32_t *factor,
+void vector_scalar_mul_aie_scalar(float *a, float *c, float *factor,
                                   int32_t N) {
   for (int i = 0; i < N; i++) {
     c[i] = *factor * a[i];
